# perf(pnp): use for loop instead of map and find in applyNodeExtensionResolution