… packages/yarnpkg-pnp/sources/hook.js