**What's the problem this PR addresses?**

The implementation of `applyNodeExtensionResolution` does a bit more work than it has to and thus ever so slightly slows down the PnP runtime.

**How did you fix it?**

- Remove unnecessary while loop since only one iteration was ever executed.
- Remove unnecessary `toFilename` calls on constants.
- Use for loop instead of map and find.

**Result**
Measured using `perf stat -r 10 -- yarn test:lint`
|before|after|
|--|--|
| 33,775824700 seconds ( +- 0,10% ) | 33,507517861 seconds ( +- 0,09% )

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I have verified that all automated PR checks pass.